requirements.txt
#certifi
#cffi
chardet
decorator
#gitdb2
GitPython
#idna
#mysqlclient ; sys_platform != 'win32'
#ext/mysqlclient-1.3.13-cp37-cp37m-win_amd64.whl ; sys_platform == 'win32'
#networkx
numpy
#pycparser
pygit2
PyGithub
#PyJWT
python-dateutil
pytz
PyYAML
regex
requests
scipy
setproctitle
six
smmap2
#SQLAlchemy
#SQLAlchemy-Utils
Unidecode
urllib3
wrapt
mpi4py
pandas
numba
